// source/components/datagridbeta/type.ts
export type Column = {
  field: string;
  headerText?: string;
  width?: number;
  filter?: boolean;
  template?: React.FC<{
    rowData: any;
    rowIndex: number;
    rowChange: (changes: any) => void;
  }>;
  tooltip?: boolean;
  showFilterPopup?: boolean;
  freezeColumn?: boolean;
  editable?: boolean;
  type?: string; // e.g., "text", "number", "date"
};